[FEAT] Add completion
