+++ src/features/game/expansion/components/potions/Intro.tsx
@@ -13,7 +13,7 @@ export const IntroPage: React.FC<IntroProps> = ({ onComplete }) => {
   return (
     <>
       <div className="p-2 pt-0 flex flex-col h-full">
-        <div className="text-[18px] leading-5 space-y-2 mb-3">
+        <div className="text-[18px] leading-5 space-y-2 mb-3 -mt-2">
           <div className="relative mt-2 float-left w-1/3">
             <DynamicNFT
               bumpkinParts={{
