+++ src/styles.css
@@ -598,6 +598,8 @@ img {
   box-sizing: border-box;
   z-index: 9999;
   transition: transform 0.5s;
+  max-width: 448px;
+  margin: 0 auto;
 }
 
 #cover {
@@ -628,4 +630,16 @@ img {
   left: 50%;
   transform: translate(-50%, -50%);
   transition: all 5s;
+}
+
+.backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.5);
+  opacity: 0;
+  transition: opacity 0.3s ease-in-out;
+  pointer-events: none;
 }
